refactor: share progress stream control lookup
